pyproject.toml
[project]
name = "rfsim-core"
version = "0.1.0"
description = "Core foundation for RF Sim"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cerberus>=1.3.7",
    "numpy>=2.2.5",
    "pint>=0.24.4",
    "pyyaml>=6.0.2",
    "scipy>=1.15.2",
    "sympy>=1.13.3",
]

[dependency-groups]
dev = [
    "pytest-benchmark>=5.1.0",
    "pytest>=8.3.5",
]

[tool.pytest.ini_options]
pythonpath = ["src"]